software/user_interface/requirements.txt
colorzero==2.0
Encoder==1.1
gpiozero==1.6.2
RPi.GPIO==0.7.1
